refactor: route node/ imports through session public API
node/ was reaching into session/internal/ directly. Re-export
CartridgeSessionAccount and signerToGuid from session/index.ts so
node/ imports from ../session instead.
