+++ packages/controller/src/session/index.ts
@@ -3,3 +3,5 @@ export * from "./provider";
 export * from "../errors";
 export * from "../types";
 export * from "./internal/types";
+export { CartridgeSessionAccount } from "./internal/account";
+export { signerToGuid } from "./internal/shared";
